Jsonwebtoken lisätty jotta vain kirjautunut kauppa voi lisätä/muokata tuotteitaan, kaupan login ominaisuus lisätty ja siihen liittyvä reducer

+++ controllers/products.js
@@ -1,6 +1,7 @@
 const productsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Product = require('../models/productModel')
-//const Shop = require('../models/shopModel')
+const Shop = require('../models/shopModel')
 
 productsRouter.get('/', async (req, res, next) => {
     try {
@@ -13,19 +14,27 @@ productsRouter.get('/', async (req, res, next) => {
 })
 
 productsRouter.post('/', async (req, res, next) => {
+    const body = req.body
     try {
-        const body = req.body
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        if (!req.token || !decodedToken.id) {
+            return res.status(401).json({ error: 'token missing or invalid' })
+        }
 
-        //const shop = await Shop.findById(id)
+        const shop = await Shop.findById(decodedToken.id)
 
         const product = new Products({
             price: body.price,
             description: body.description,
             availability: true,
-            // shop: shop._id
+            shop: shop._id
         })
 
         const savedProduct = await product.save()
+
+        shop.products = shop.products.concat(savedProduct._id)
+        await shop.save()
+
         res.json(savedProduct.toJSON())
 
     } catch (exception) {
@@ -34,16 +43,17 @@ productsRouter.post('/', async (req, res, next) => {
 })
 
 productsRouter.put('/:id', async (req, res, next) => {
+    const body = req.body
     try {
-        const body = req.body
-
-        //const shop = await Shop.findById(id)
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        if (!req.token || !decodedToken.id) {
+            return res.status(401).json({ error: 'token missing or invalid' })
+        }
 
         const product = {
             price: body.price,
             description: body.description,
-            availability: body.availability,
-            // shop: shop._id
+            availability: body.availability
         }
 
         const modifiedProduct = await Product.findByIdAndUpdate(req.params.id, product, { new: true })
@@ -56,9 +66,19 @@ productsRouter.put('/:id', async (req, res, next) => {
 
 productsRouter.delete('/:id', async (req, res, next) => {
     try {
-        await Product.findByIdAndRemove(req.params.id)
-        res.status(204).end()
+        const product = await Product.findById(req.params.id)
 
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+        const shop = await Shop.findById(decodedToken.id)
+
+        if (product.shop.toString() === shop.id.toString()) {
+            await Product.findByIdAndRemove(req.params.id)
+            res.status(204).end()
+        } else {
+            return res.status(401).json({ error: 'only logged in shop can delete products' })
+        }
+        
     } catch (exception) {
         next(exception)
     }
